bot(aiogramm)/bot_current/requirements.txt
python-dotenv==1.0.1
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asyncmy==0.2.9
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
fastapi==0.115.5
fonttools==4.55.0
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
kiwisolver==1.4.7
magic-filter==1.0.12
matplotlib==3.9.2
multidict==6.1.0
numpy==2.1.3
packaging==24.2
pillow==11.0.0
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
PyMySQL==1.1.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.18.0

